requirements.txt
numpy==1.20.1
six==1.15.0
websocket-client==0.58.0
